requirements.txt
aiogram>=3.4.1
httpx>=0.27.0
pydantic>=2.5
websockets>=12.0
certifi>=2024.2.2
python-dotenv>=1.0.1